+++ README.md
@@ -17,5 +17,6 @@
 ### 2. 리팩토링 및 테스트
 - [x] 회원가입/로그인 로직 리팩토링
 - [x] 회원가입/로그인 로직 테스트 
+- [x] 디렉토리 구조 변경
 - [ ] 사용자의 위시리스트에 상품 추가, 수정, 삭제, 조회하는 로직 리팩토링
