--- src/main/java/gift/GlobalExceptionHandler.java
@@ -1,52 +1,0 @@
-package gift;
-
-import gift.Login.common.exception.UserNotFoundException;
-import gift.product.exception.ProductAlreadyExistsException;
-import gift.product.exception.ProductNotFoundException;
-import jakarta.validation.ConstraintViolationException;
-import org.springframework.http.HttpStatus;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@ControllerAdvice
-public class GlobalExceptionHandler {
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    @ExceptionHandler(ProductNotFoundException.class)
-    public String handleProductNotFoundException(ProductNotFoundException ex, Model model) {
-        model.addAttribute("errorMessage", ex.getMessage());
-        return "ErrorPage/NotFound";
-    }
-
-    @ResponseStatus(HttpStatus.CONFLICT)
-    @ExceptionHandler(ProductAlreadyExistsException.class)
-    public String handleProductAlreadyExistsException(ProductAlreadyExistsException ex, Model model) {
-        model.addAttribute("errorMessage", ex.getMessage());
-        return "ErrorPage/AlreadyExists";
-    }
-
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(ConstraintViolationException.class)
-    public String handleConstraintViolationException(ConstraintViolationException ex, Model model) {
-        Map<String, String> errors = new HashMap<>();
-        ex.getConstraintViolations().forEach(cv -> {
-            String fieldName = cv.getPropertyPath().toString();
-            String errorMessage = cv.getMessage();
-            errors.put(fieldName, errorMessage);
-        });
-        model.addAttribute("validationErrors", errors);
-        return "ErrorPage/BadRequest";
-    }
-
-    @ExceptionHandler(UserNotFoundException.class)
-    @ResponseStatus(HttpStatus.UNAUTHORIZED)
-    @ResponseBody
-    public String handleUserNotFoundException(UserNotFoundException ex) {
-        return ex.getMessage();
-    }
-}
