--- src/main/java/gift/Login/common/auth/LoginMember.java
@@ -1,12 +1,0 @@
-package gift.Login.common.auth;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface LoginMember {
-}
-
