--- src/main/java/gift/Login/common/config/WebConfig.java
@@ -1,28 +1,0 @@
-package gift.Login.common.config;
-
-import gift.Login.common.auth.JwtUtil;
-import gift.Login.common.auth.LoginMemberArgumentResolver;
-import gift.Login.repository.MemberRepository;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.List;
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-
-    private final JwtUtil jwtUtil;
-    private final MemberRepository memberRepository;
-
-    public WebConfig(JwtUtil jwtUtil, MemberRepository memberRepository) {
-        this.jwtUtil = jwtUtil;
-        this.memberRepository = memberRepository;
-    }
-
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(new LoginMemberArgumentResolver(jwtUtil, memberRepository));
-    }
-}
