--- src/main/java/gift/Login/common/exception/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package gift.Login.common.exception;
-
-public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(String message) {
-        super(message);
-    }
-}
