--- src/main/java/gift/Login/model/Member.java
@@ -1,74 +1,0 @@
-package gift.Login.model;
-
-import jakarta.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Table(name = "member")
-public class Member {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    @Column(nullable = false)
-    private String password;
-
-    @OneToMany(mappedBy = "member", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.LAZY)
-    private List<Wish> wishList = new ArrayList<>();
-
-    // Constructors, Getters, and Setters
-    public Member() {}
-
-    public Member(String email, String password) {
-        this.email = email;
-        this.password = password;
-    }
-
-    // Getters and setters
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public List<Wish> getWishList() {
-        return wishList;
-    }
-
-    public void setWishList(List<Wish> wishList) {
-        this.wishList = wishList;
-    }
-
-    public void addWish(Wish wish) {
-        wishList.add(wish);
-        wish.setMember(this);
-    }
-
-    public void removeWish(Wish wish) {
-        wishList.remove(wish);
-        wish.setMember(null);
-    }
-}
