--- src/main/java/gift/Login/model/ResponseToken.java
@@ -1,18 +1,0 @@
-package gift.Login.model;
-
-public class ResponseToken {
-    private String token;
-
-    public ResponseToken(String token) {
-        this.token = token;
-    }
-
-    public String getToken() {
-        return token;
-    }
-
-    public void setToken(String token) {
-        this.token = token;
-    }
-}
-
