--- src/main/java/gift/Login/model/Wish.java
@@ -1,53 +1,0 @@
-package gift.Login.model;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "wish")
-public class Wish {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "product_id", nullable = false)
-    private Product product;
-
-    // Constructors, Getters, and Setters
-    public Wish() {}
-
-    public Wish(Member member, Product product) {
-        this.member = member;
-        this.product = product;
-    }
-
-    // Getters and setters
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public Member getMember() {
-        return member;
-    }
-
-    public void setMember(Member member) {
-        this.member = member;
-    }
-
-    public Product getProduct() {
-        return product;
-    }
-
-    public void setProduct(Product product) {
-        this.product = product;
-    }
-}
