--- src/main/java/gift/Login/repository/MemberRepository.java
@@ -1,12 +1,0 @@
-package gift.Login.repository;
-
-import gift.Login.model.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByEmail(String email);
-}
