--- src/main/java/gift/Login/repository/ProductRepository.java
@@ -1,9 +1,0 @@
-package gift.Login.repository;
-
-import gift.Login.model.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
-}
