+++ src/main/java/gift/common/auth/JwtUtil.java
@@ -1,0 +1,49 @@
+package gift.common.auth;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+import java.nio.charset.StandardCharsets;
+import java.security.Key;
+
+@Component
+public class JwtUtil {
+
+    private final Key key;
+
+    public JwtUtil(@Value("${jwt.secret}") String secretKey) {
+        this.key = Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
+    }
+
+    // JWT 토큰을 파싱하여 그 안에 포함된 클레임을 반환
+    public Claims extractClaims(String token) {
+        return Jwts.parser()
+                .setSigningKey(key)
+                .build()
+                .parseClaimsJws(token)
+                .getBody();
+    }
+
+    // 토큰이 유효한지 확인
+    public boolean isTokenValid(String token) {
+        try {
+            Jwts.parser()
+                    .setSigningKey(key)
+                    .build()
+                    .parseClaimsJws(token);
+            return true;
+        } catch (Exception e) {
+            e.printStackTrace();
+            return false;
+        }
+    }
+
+    // 토큰에서 이메일 주소 추출
+    public String extractEmail(String token) {
+        Claims claims = extractClaims(token);
+        return claims.get("email", String.class);
+    }
+}
