+++ src/main/java/gift/common/config/WebConfig.java
@@ -1,0 +1,28 @@
+package gift.common.config;
+
+import gift.common.auth.JwtUtil;
+import gift.common.auth.LoginMemberArgumentResolver;
+import gift.member.repository.MemberRepository;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import java.util.List;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+    private final JwtUtil jwtUtil;
+    private final MemberRepository memberRepository;
+
+    public WebConfig(JwtUtil jwtUtil, MemberRepository memberRepository) {
+        this.jwtUtil = jwtUtil;
+        this.memberRepository = memberRepository;
+    }
+
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(new LoginMemberArgumentResolver(jwtUtil, memberRepository));
+    }
+}
