+++ src/main/java/gift/common/exception/UserNotFoundException.java
@@ -1,0 +1,7 @@
+package gift.common.exception;
+
+public class UserNotFoundException extends RuntimeException {
+    public UserNotFoundException(String message) {
+        super(message);
+    }
+}
