+++ src/main/java/gift/member/repository/MemberRepository.java
@@ -1,0 +1,12 @@
+package gift.member.repository;
+
+import gift.member.model.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    Optional<Member> findByEmail(String email);
+}
