--- src/main/java/gift/product/exception/InvalidProductDataException.java
@@ -1,7 +1,0 @@
-package gift.product.exception;
-
-public class InvalidProductDataException extends RuntimeException {
-    public InvalidProductDataException(String value) {
-        super(value + "을(를) 다시 입력해주세요.");
-    }
-}
