--- src/main/java/gift/product/exception/ProductAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package gift.product.exception;
-
-public class ProductAlreadyExistsException extends RuntimeException {
-    public ProductAlreadyExistsException(String name) {
-        super("상품 :" + name + "는 이미 존재하는 상품입니다." );
-    }
-}
