--- src/main/java/gift/product/exception/ProductNotFoundException.java
@@ -1,7 +1,0 @@
-package gift.product.exception;
-
-public class ProductNotFoundException extends RuntimeException {
-    public ProductNotFoundException(Long id) {
-        super(id + "에 해당하는 상품이 존재하지 않습니다.");
-    }
-}
