+++ src/main/java/gift/product/repository/ProductRepository.java
@@ -1,14 +1,9 @@
 package gift.product.repository;
 
 import gift.product.model.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
-import java.util.List;
-import java.util.Optional;
-
-public interface ProductRepository {
-    List<Product> findAll();
-    Optional<Product> findById(Long id);
-    Product save(Product product);
-    Product update(Product product);
-    void deleteById(Long id);
+@Repository
+public interface ProductRepository extends JpaRepository<Product, Long> {
 }
